clean(core): Move IPFSResource into new package dev.enola.cas
