+++ java/dev/enola/cas/IPFSResource.java
@@ -15,11 +15,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package dev.enola.common.io.resource;
+package dev.enola.cas;
 
 import com.google.common.base.Strings;
 import com.google.common.io.ByteSource;
 import com.google.common.net.MediaType;
+
+import dev.enola.common.io.resource.*;
 
 import io.ipfs.cid.Cid;
